+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improve documentation of components by adding component versions
 
+### Fixed
+
+- Change color of indicator and borders in high contrast mode (light scheme) for radio and checkbox components (#645)
+
 ## [0.15.0](https://github.com/Orange-OpenSource/ouds-ios/compare/0.14.0...0.15.0) - 2025-05-28
 
 ### Added